exceptions handler

+++ src/main/java/com/xicaum59/desafio3/services/ClientService.java
@@ -11,6 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 import com.xicaum59.desafio3.DTO.ClientDTO;
 import com.xicaum59.desafio3.entities.Client;
 import com.xicaum59.desafio3.epositories.ClientRepository;
+import com.xicaum59.desafio3.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -22,7 +25,8 @@ public class ClientService {
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
 		Optional<Client> result = repository.findById(id);
-		Client client = result.get();
+		
+		Client client = result.orElseThrow(() -> new ResourceNotFoundException("Resource not found"));
 		return new ClientDTO(client);
 		
 	}
@@ -45,14 +49,25 @@ public class ClientService {
 	
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
-		Client client = repository.getReferenceById(id);
-		copyDtoToEntity(dto, client);
-		client = repository.save(client);
-		return new ClientDTO(client);
+		
+		try {
+			Client client = repository.getReferenceById(id);
+			copyDtoToEntity(dto, client);
+			client = repository.save(client);
+			return new ClientDTO(client);
+		}
+		catch (EntityNotFoundException e){
+			throw new ResourceNotFoundException("Resource not found");
+			
+		}
 	}
 	
 	@Transactional
-	public void delete(Long id) {
+	public void delete(Long id) {		
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Resource not found");
+		}
+		
 		repository.deleteById(id);
 		
 	}
